agents/logistics-agent/requirements.txt
a2a-sdk[http-server]==0.3.2
uvicorn
starlette
python-dotenv
